+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Version 0.6
 
+* Simplified parsing to use `vector<string>` only
 * Fixed fallthrough, made it optional as well (default: off): `.fallthrough()`.
 * Added string versions of `->requires()` and `->excludes()` for consistency.
 * Renamed protected members for internal consistency, grouped docs.